+++ cli/account/help.py
@@ -1,0 +1,4 @@
+"""
+Provide help messages for command line interface's account commands.
+"""
+GET_ACCOUNT_BALANCE_ADDRESS_ARGUMENT_HELP_MESSAGE = 'Get balance of the account by its address.'
