+++ cli/account/interfaces.py
@@ -1,0 +1,15 @@
+"""
+Provide implementation of the account interfaces.
+"""
+
+
+class AccountInterface:
+    """
+    Implements account interface.
+    """
+
+    async def get_balance(self, address):
+        """
+        Get balance of the account by its address.
+        """
+        pass
