+++ cli/constants.py
@@ -1,0 +1,8 @@
+"""
+Provide constants for command line interface.
+"""
+ADDRESS_REGEXP = '[0-9a-f]{70}'
+
+FAILED_EXIT_FROM_COMMAND = -1
+
+NODE_URL_ARGUMENT_HELP_MESSAGE = 'Apply the command to the specified node by its URL.'
